# swagger.json
{
  "basePath": "/4537/termproject/API/V1",
  "consumes":[],
...
  "host": "michealozdoba.com",
  "info": {
    "version": "1.0.0",
    "title": "COMP4537 Term Project API",
    "description": "Open API Specifications for COMP4537 Term Project",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
    "/post": {
      "delete": {
        "consumes": ["application/x-www-form-urlencoded"],
        "produces":["application/json"],
        "deprecated": false,
        "description": "Delete a Post",
        "operationId": "deletePostById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "ID of post to delete",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted post with id"
          },
          "400": {
            "description": "Failed to delete post with id"
          }
        },
        "summary": "Delete a location",
        "tags": ["post"]
      },
      "get": {
        "consumes": [],
        "deprecated": false,
        "description": "Returns all locations",
        "externalDocs": {},
        "operationId": "selectAllFromLocation",
        "parameters": [],
        "produces":["text/html"],
        "responses": {
          "200": {
            "description": "Successful Response"
          }
        },
        "security": [],
        "schemes": [],
        "summary": "Return a list of all locations",
        "tags": ["post"]
      },
      "post": {
        "consumes": [
          "application/x-www-form-urlencoded",
          "text/html"
        ],
        "deprecated": false,
        "description": "Create a location",
        "operationId": "createLocationByName",
        "parameters": [
          {
            "name": "location",
            "in": "body",
            "description": "Name of location to create",
            "required": true,
            "schema": {
              "type": "object",
              "example": {
                "location": "Kelowna"
              }
            }
          }
        ],
        "produces":["text/html"],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Create a location",
        "tags": ["post"]
      },
      "put": {
        "consumes": [
          "application/x-www-form-urlencoded",
          "text/html"
        ],
        "deprecated": false,
        "description": "Create a location",
        "operationId": "createLocationByName",
        "parameters": [
          {
            "name": "location",
            "in": "body",
            "description": "Name of location to create",
            "required": true,
            "schema": {
              "type": "object",
              "example": {
                "location": "Kelowna"
              }
            }
          }
        ],
        "produces":["text/html"],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "summary": "Create a location",
        "tags": ["post"]
      }
    },
...
  },
...
  "produces":[],
...
  "schemes": ["https"],
  "security": [],
  "securityDefinitions": {},
  "swagger": "2.0",
...
}